Add GetStarted screen and BasicAppButton widget, update minSdk and AppAssets and AppVector classes
The commit adds a new screen called GetStarted and a new widget called BasicAppButton. It also updates the minSdk version to 23 and adds new assets to the AppAssets and AppVector classes. The commit also includes changes to the main.dart file to remove the debug banner and set the home screen to the SplashPage.